+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 0.9.10
 
+- Add new staff: Diamond Holy Staff
 - Add particle effects on Divine Protection pop
 - Update render effect API usage
-
+- 
 # 0.9.9
 
 - Add new spell for Holy Wands: Heal